openai embed instrumentation
