+++ src/instrumentation/vercel/instrumentation.ts
@@ -3,7 +3,7 @@ import { diag } from '@opentelemetry/api'
 import { InstrumentationBase, InstrumentationModuleDefinition, InstrumentationNodeModuleDefinition } from '@opentelemetry/instrumentation'
 import { version, name } from '../../../package.json'
 import { APIS } from '@langtrase/trace-attributes'
-import { generateTextPatch } from './patch'
+import { embedPatch} from '@langtrace-instrumentation/vercel/patch'
 
 class VercelAIInstrumentation extends InstrumentationBase<any> {
   constructor () {
@@ -35,10 +35,14 @@ class VercelAIInstrumentation extends InstrumentationBase<any> {
 
   public _patch (moduleExports: any, moduleVersion?: string): any {
     this._wrap(moduleExports, 'generateText', (original: any) => generateTextPatch.call(this, original as (...args: any[]) => any, APIS.ai.GENERATE_TEXT.METHOD, this.tracer, version, name, moduleVersion))
+    this._wrap(moduleExports, 'embed', (original: any) => embedPatch.call(this, original as (...args: any[]) => any, APIS.ai.EMBED.METHOD, this.tracer, version, name, moduleVersion))
+    this._wrap(moduleExports, 'embedMany', (original: any) => embedPatch.call(this, original as (...args: any[]) => any, APIS.ai.EMBED_MANY.METHOD, this.tracer, version, name, moduleVersion))
   }
 
   public _unpatch (module: any): void {
     this._unwrap(module, 'generateText')
+    this._unwrap(module, 'embed')
+    this._unwrap(module, 'embedMany')
   }
 }
 export const vercelAIInstrumentation = new VercelAIInstrumentation()
